ui: cart screen added
