+++ example/lib/routes/routes.dart
@@ -1,3 +1,4 @@
+import 'package:example/views/cart_view.dart';
 import 'package:example/views/products_view.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
@@ -5,10 +6,17 @@ import 'package:go_router/go_router.dart';
 final GoRouter router = GoRouter(
   routes: <RouteBase>[
     GoRoute(
-      path: '/',
-      builder: (BuildContext context, GoRouterState state) {
-        return const ProductsView();
-      },
-    ),
+        path: '/',
+        builder: (BuildContext context, GoRouterState state) {
+          return const ProductsView();
+        },
+        routes: [
+          GoRoute(
+            path: 'cart',
+            builder: (BuildContext context, GoRouterState state) {
+              return const CartView();
+            },
+          ),
+        ]),
   ],
 );
